+++ test/configuration_test.dart
@@ -5,7 +5,6 @@
 //
 // Requirements: the "test/configuration_test.yaml" file
 
-import 'dart:io';
 import 'dart:math';
 import 'package:test/test.dart';
 import 'package:dartdap/dartdap.dart';
@@ -131,7 +130,7 @@ void main() {
         await ldap_conf.getConnection();
         expect(false, isTrue, reason: "Unreachable");
       } catch (e) {
-        expect(e, new isInstanceOf<LDAPException>());
+        expect(e, new isInstanceOf<LdapException>());
       }
     });
 
@@ -141,7 +140,7 @@ void main() {
         await ldap_conf.getConnection();
         expect(false, isTrue, reason: "Unreachable");
       } catch (e) {
-        expect(e, new isInstanceOf<LDAPException>());
+        expect(e, new isInstanceOf<LdapException>());
       }
     });
 
@@ -151,7 +150,7 @@ void main() {
         await ldap_conf.getConnection();
         expect(false, isTrue, reason: "Unreachable");
       } catch (e) {
-        expect(e, new isInstanceOf<LDAPException>());
+        expect(e, new isInstanceOf<LdapException>());
       }
     });
 
@@ -164,7 +163,7 @@ void main() {
         await ldap_conf.getConnection();
         expect(false, isTrue, reason: "Unreachable");
       } catch (e) {
-        expect(e, new isInstanceOf<LDAPException>());
+        expect(e, new isInstanceOf<LdapException>());
       }
     });
 
@@ -179,7 +178,7 @@ void main() {
         expect(ldap_conf.ssl, equals(false)); // default to no TLS
         expect(ldap_conf.bindDN, equals("")); // default to no bindDN
         expect(ldap_conf.password, equals("")); // default to no password
-        expect(e, new isInstanceOf<SocketException>()); // but could not connect
+        expect(e, new isInstanceOf<LdapSocketServerNotFoundException>()); // but could not connect
       }
     });
 
@@ -192,7 +191,7 @@ void main() {
         await ldap_conf.getConnection();
         expect(false, isTrue, reason: "Unreachable");
       } catch (e) {
-        expect(e, new isInstanceOf<LDAPException>());
+        expect(e, new isInstanceOf<LdapException>());
       }
     });
 
@@ -207,7 +206,7 @@ void main() {
         expect(ldap_conf.ssl, equals(false)); // default to no TLS
         expect(ldap_conf.bindDN, equals("")); // default to no bindDN
         expect(ldap_conf.password, equals("")); // default to no password
-        expect(e, new isInstanceOf<SocketException>()); // but could not connect
+        expect(e, new isInstanceOf<LdapSocketServerNotFoundException>()); // but could not connect
       }
     });
 
@@ -220,7 +219,7 @@ void main() {
         await ldap_conf.getConnection();
         expect(false, isTrue, reason: "Unreachable");
       } catch (e) {
-        expect(e, new isInstanceOf<LDAPException>());
+        expect(e, new isInstanceOf<LdapException>());
       }
     });
 
@@ -236,7 +235,7 @@ void main() {
         expect(ldap_conf.ssl, equals(false)); // ssl loaded
         expect(ldap_conf.bindDN, equals("")); // default to no bindDN
         expect(ldap_conf.password, equals("")); // default to no password
-        expect(e, new isInstanceOf<SocketException>()); // but could not connect
+        expect(e, new isInstanceOf<LdapSocketServerNotFoundException>()); // but could not connect
       }
     });
 
@@ -252,7 +251,7 @@ void main() {
         expect(ldap_conf.ssl, equals(true)); // ssl loaded
         expect(ldap_conf.bindDN, equals("")); // default to no bindDN
         expect(ldap_conf.password, equals("")); // default to no password
-        expect(e, new isInstanceOf<SocketException>()); // but could not connect
+        expect(e, new isInstanceOf<LdapSocketServerNotFoundException>()); // but could not connect
       }
     });
 
@@ -268,7 +267,7 @@ void main() {
         expect(ldap_conf.ssl, equals(false)); // ssl loaded
         expect(ldap_conf.bindDN, equals("")); // default to no bindDN
         expect(ldap_conf.password, equals("")); // default to no password
-        expect(e, new isInstanceOf<SocketException>()); // but could not connect
+        expect(e, new isInstanceOf<LdapSocketServerNotFoundException>()); // but could not connect
       }
     });
 
@@ -284,7 +283,7 @@ void main() {
         expect(ldap_conf.ssl, equals(true)); // ssl value was loaded
         expect(ldap_conf.bindDN, equals("")); // default to no bindDN
         expect(ldap_conf.password, equals("")); // default to no password
-        expect(e, new isInstanceOf<SocketException>()); // but could not connect
+        expect(e, new isInstanceOf<LdapSocketServerNotFoundException>()); // but could not connect
       }
     });
 
@@ -302,7 +301,7 @@ void main() {
         expect(ldap_conf.ssl, equals(false)); // ssl value was loaded
         expect(ldap_conf.bindDN, equals("dc=example,dc=com")); // bindDN loaded
         expect(ldap_conf.password, equals("p@ssw0rd")); // password loaded
-        expect(e, new isInstanceOf<SocketException>()); // but could not connect
+        expect(e, new isInstanceOf<LdapSocketServerNotFoundException>()); // but could not connect
       }
     });
   });